# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EduTrack - Tutor & Coach Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DashboardSummary {
  totalStudents: number;
  classesToday: number;
  attendanceRate: number;
  pendingBillings: number;
}

export interface ClassPerformance {
  classId: number;
  className: string;
  averageScore: number;
  color: string;
}

export interface TodayCheckin {
  id: number;
  studentId: number;
  studentName: string;
  className: string;
  scheduledTime: string;
  status: string;
  /** @nullable */
  checkinTime?: string | null;
  /** @nullable */
  checkoutTime?: string | null;
}

export interface Student {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  grade?: string | null;
  enrolledAt: string;
  status: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateStudentBody {
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  grade?: string | null;
  status?: string;
}

export interface UpdateStudentBody {
  name?: string;
  email?: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  grade?: string | null;
  status?: string;
}

export interface Teacher {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  role: string;
  subjects: string;
  status: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateTeacherBody {
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  role?: string;
  subjects: string;
  status?: string;
}

export interface UpdateTeacherBody {
  name?: string;
  email?: string;
  /** @nullable */
  phone?: string | null;
  role?: string;
  subjects?: string;
  status?: string;
}

export interface Class {
  id: number;
  name: string;
  subject: string;
  /** @nullable */
  teacherId?: number | null;
  /** @nullable */
  teacherName?: string | null;
  capacity: number;
  enrolledCount: number;
  status: string;
  /** @nullable */
  description?: string | null;
  color: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateClassBody {
  name: string;
  subject: string;
  /** @nullable */
  teacherId?: number | null;
  capacity: number;
  /** @nullable */
  description?: string | null;
  color?: string;
  status?: string;
}

export interface UpdateClassBody {
  name?: string;
  subject?: string;
  /** @nullable */
  teacherId?: number | null;
  capacity?: number;
  /** @nullable */
  description?: string | null;
  color?: string;
  status?: string;
}

export interface Checkin {
  id: number;
  studentId: number;
  studentName: string;
  classId: number;
  className: string;
  date: string;
  scheduledTime: string;
  status: string;
  /** @nullable */
  checkinTime?: string | null;
  /** @nullable */
  checkoutTime?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateCheckinBody {
  studentId: number;
  classId: number;
  date: string;
  scheduledTime: string;
  status?: string;
  /** @nullable */
  notes?: string | null;
}

export interface UpdateCheckinBody {
  status?: string;
  /** @nullable */
  checkinTime?: string | null;
  /** @nullable */
  checkoutTime?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface Assessment {
  id: number;
  studentId: number;
  studentName: string;
  classId: number;
  className: string;
  title: string;
  score: number;
  maxScore: number;
  percentage: number;
  date: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateAssessmentBody {
  studentId: number;
  classId: number;
  title: string;
  score: number;
  maxScore: number;
  date: string;
  /** @nullable */
  notes?: string | null;
}

export interface UpdateAssessmentBody {
  title?: string;
  score?: number;
  maxScore?: number;
  date?: string;
  /** @nullable */
  notes?: string | null;
}

export interface BillingRecord {
  id: number;
  studentId: number;
  studentName: string;
  /** @nullable */
  classId?: number | null;
  /** @nullable */
  className?: string | null;
  amount: number;
  status: string;
  reason: string;
  /** @nullable */
  dueDate?: string | null;
  /** @nullable */
  paidDate?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateBillingRecordBody {
  studentId: number;
  /** @nullable */
  classId?: number | null;
  amount: number;
  status?: string;
  reason: string;
  /** @nullable */
  dueDate?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface UpdateBillingRecordBody {
  amount?: number;
  status?: string;
  reason?: string;
  /** @nullable */
  dueDate?: string | null;
  /** @nullable */
  paidDate?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface ScheduleSlot {
  id: number;
  classId: number;
  className: string;
  /** @nullable */
  teacherName?: string | null;
  dayOfWeek: string;
  startTime: string;
  endTime: string;
  /** @nullable */
  room?: string | null;
  color: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateScheduleSlotBody {
  classId: number;
  dayOfWeek: string;
  startTime: string;
  endTime: string;
  /** @nullable */
  room?: string | null;
}

export interface UpdateScheduleSlotBody {
  classId?: number;
  dayOfWeek?: string;
  startTime?: string;
  endTime?: string;
  /** @nullable */
  room?: string | null;
}

export type ListCheckinsParams = {
  date?: string;
};