# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EduTrack - Tutor & Coach Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /classes:
    get:
      operationId: listClasses
      tags: [classes]
      summary: List all classes
      responses:
        "200":
          description: List of classes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Class"
    post:
      operationId: createClass
      tags: [classes]
      summary: Create a class
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClassBody"
      responses:
        "201":
          description: Created class
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
  /classes/{id}:
    get:
      operationId: getClass
      tags: [classes]
      summary: Get a class by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Class
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
        "404":
          description: Not found
    patch:
      operationId: updateClass
      tags: [classes]
      summary: Update a class
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClassBody"
      responses:
        "200":
          description: Updated class
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
        "404":
          description: Not found
    delete:
      operationId: deleteClass
      tags: [classes]
      summary: Delete a class
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Check-ins ──
  /checkins:
    get:
      operationId: listCheckins
      tags: [checkins]
      summary: List check-ins
      parameters:
        - name: date
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of check-ins
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Checkin"
    post:
      operationId: createCheckin
      tags: [checkins]
      summary: Create a check-in record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCheckinBody"
      responses:
        "201":
          description: Created check-in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Checkin"
  /checkins/{id}:
    patch:
      operationId: updateCheckin
      tags: [checkins]
      summary: Update a check-in (e.g. check-out, change status)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCheckinBody"
      responses:
        "200":
          description: Updated check-in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Checkin"
        "404":
          description: Not found
    delete:
      operationId: deleteCheckin
      tags: [checkins]
      summary: Delete a check-in
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Assessments ──
  /assessments:
    get:
      operationId: listAssessments
      tags: [assessments]
      summary: List assessments
      responses:
        "200":
          description: List of assessments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Assessment"
    post:
      operationId: createAssessment
      tags: [assessments]
      summary: Create an assessment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssessmentBody"
      responses:
        "201":
          description: Created assessment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assessment"
  /assessments/{id}:
    patch:
      operationId: updateAssessment
      tags: [assessments]
      summary: Update an assessment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAssessmentBody"
      responses:
        "200":
          description: Updated assessment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assessment"
        "404":
          description: Not found
    delete:
      operationId: deleteAssessment
      tags: [assessments]
      summary: Delete an assessment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Billing ──
  /billing:
    get:
      operationId: listBillingRecords
      tags: [billing]
      summary: List billing records
      responses:
        "200":
          description: List of billing records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BillingRecord"
    post:
      operationId: createBillingRecord
      tags: [billing]
      summary: Create a billing record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBillingRecordBody"
      responses:
        "201":
          description: Created billing record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillingRecord"
  /billing/{id}:
    patch:
      operationId: updateBillingRecord
      tags: [billing]
      summary: Update a billing record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBillingRecordBody"
      responses:
        "200":
          description: Updated billing record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillingRecord"
        "404":
          description: Not found
    delete:
      operationId: deleteBillingRecord
      tags: [billing]
      summary: Delete a billing record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Schedule ──
...
components:
  schemas:
...
    Class:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        subject:
          type: string
        teacherId:
          type: ["integer", "null"]
        teacherName:
          type: ["string", "null"]
        capacity:
          type: integer
        enrolledCount:
          type: integer
        status:
          type: string
        description:
          type: ["string", "null"]
        color:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - subject
        - capacity
        - enrolledCount
        - status
        - color
        - createdAt
        - updatedAt
    CreateClassBody:
      type: object
      properties:
        name:
          type: string
        subject:
          type: string
        teacherId:
          type: ["integer", "null"]
        capacity:
          type: integer
        description:
          type: ["string", "null"]
        color:
          type: string
          default: blue
        status:
          type: string
          default: active
      required:
        - name
        - subject
        - capacity
    UpdateClassBody:
      type: object
      properties:
        name:
          type: string
        subject:
          type: string
        teacherId:
          type: ["integer", "null"]
        capacity:
          type: integer
        description:
          type: ["string", "null"]
        color:
          type: string
        status:
          type: string
    Checkin:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
        classId:
          type: integer
        className:
          type: string
        date:
          type: string
        scheduledTime:
          type: string
        status:
          type: string
        checkinTime:
          type: ["string", "null"]
        checkoutTime:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - studentId
        - studentName
        - classId
        - className
        - date
        - scheduledTime
        - status
        - createdAt
        - updatedAt
    CreateCheckinBody:
      type: object
      properties:
        studentId:
          type: integer
        classId:
          type: integer
        date:
          type: string
        scheduledTime:
          type: string
        status:
          type: string
          default: pending
        notes:
          type: ["string", "null"]
      required:
        - studentId
        - classId
        - date
        - scheduledTime
    UpdateCheckinBody:
      type: object
      properties:
        status:
          type: string
        checkinTime:
          type: ["string", "null"]
        checkoutTime:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    Assessment:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
        classId:
          type: integer
        className:
          type: string
        title:
          type: string
        score:
          type: number
        maxScore:
          type: number
        percentage:
          type: number
        date:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - studentId
        - studentName
        - classId
        - className
        - title
        - score
        - maxScore
        - percentage
        - date
        - createdAt
        - updatedAt
    CreateAssessmentBody:
      type: object
      properties:
        studentId:
          type: integer
        classId:
          type: integer
        title:
          type: string
        score:
          type: number
        maxScore:
          type: number
        date:
          type: string
        notes:
          type: ["string", "null"]
      required:
        - studentId
        - classId
        - title
        - score
        - maxScore
        - date
    UpdateAssessmentBody:
      type: object
      properties:
        title:
          type: string
        score:
          type: number
        maxScore:
          type: number
        date:
          type: string
        notes:
          type: ["string", "null"]
    BillingRecord:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
        classId:
          type: ["integer", "null"]
        className:
          type: ["string", "null"]
        amount:
          type: number
        status:
          type: string
        reason:
          type: string
        dueDate:
          type: ["string", "null"]
        paidDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - studentId
        - studentName
        - amount
        - status
        - reason
        - createdAt
        - updatedAt
    CreateBillingRecordBody:
      type: object
      properties:
        studentId:
          type: integer
        classId:
          type: ["integer", "null"]
        amount:
          type: number
        status:
          type: string
          default: pending
        reason:
          type: string
        dueDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required:
        - studentId
        - amount
        - reason
    UpdateBillingRecordBody:
      type: object
      properties:
        amount:
          type: number
        status:
          type: string
        reason:
          type: string
        dueDate:
          type: ["string", "null"]
        paidDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
...