# lib/api-zod/src/generated/types/checkin.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EduTrack - Tutor & Coach Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Checkin {
  id: number;
  studentId: number;
  studentName: string;
  classId: number;
  className: string;
  date: string;
  scheduledTime: string;
  status: string;
  /** @nullable */
  checkinTime?: string | null;
  /** @nullable */
  checkoutTime?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}