# lib/api-zod/src/generated/types/createCheckinBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EduTrack - Tutor & Coach Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCheckinBody {
  studentId: number;
  classId: number;
  date: string;
  scheduledTime: string;
  status?: string;
  /** @nullable */
  notes?: string | null;
}