# lib/api-zod/src/generated/types/updateAssessmentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EduTrack - Tutor & Coach Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateAssessmentBody {
  title?: string;
  score?: number;
  maxScore?: number;
  date?: string;
  /** @nullable */
  notes?: string | null;
}