# lib/api-zod/src/generated/types/updateClassBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EduTrack - Tutor & Coach Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateClassBody {
  name?: string;
  subject?: string;
  /** @nullable */
  teacherId?: number | null;
  capacity?: number;
  /** @nullable */
  description?: string | null;
  color?: string;
  status?: string;
}